command file structure
