+++ README.md
@@ -1,9 +1,21 @@
 # Markblog 📖
 The way to blog with markdown. 
 
-## Vision 👁‍🗨
+# TODO: 
+- [ ] build-command
+- [ ] init-command 
+- [ ] collection of posts
+- [ ] config in separate file? 
+
+## Idea 👁‍🗨
 I want to write a blog. I also want to write using [markdown](https://en.wikipedia.org/wiki/Markdown), in my own environment. 
 There are ways to do this. However, I did not find anything 
 that was as simple and "plug-and-play"-like as I wanted. 
 
 This need is what I want _Markblog_ to fix for me. 
+
+## Setup
+You only need three things: 
+1. ./index.md, for writing content of the front page. 
+2. ./posts, containing files for each post.
+3. ./style.css
