+++ commands/help.ts
@@ -1,0 +1,4 @@
+export const show_help = () => {
+
+    console.log("HELP PAGE"); //TODO
+}
