+++ markblog.ts
@@ -1,6 +1,5 @@
-import {get_posts} from "./posts.ts";
-import {write_file, assemble_html_page} from "./common.ts"
-import {get_index} from "./index.ts";
+import { build } from "./commands/build.ts";
+import { show_help } from "./commands/help.ts";
 
 const options = {
     post_source: "./posts", 
@@ -10,23 +9,24 @@ const options = {
 }
 
 
-try {
-    await Deno.mkdir(options.post_destination);
-} catch(error) {// already exists
+switch(Deno.args[0]?.toLowerCase()) {
+
+    case "build": 
+        await build(options);
+        break;
+    case "init": 
+        throw "init is not implemented";
+        break;
+    default: 
+        show_help();
+        break;
 }
 
-const posts = await get_posts(options.post_source, options.post_destination);
-const index = await get_index(posts);
 
-posts.forEach(async post => {
 
-    const html = await assemble_html_page(post.html, options.post_style)
-    write_file(post.location, html);
-});
+const init = () => {
 
-const content = index.main_content.concat(index.links);
-const html = await assemble_html_page(content, options.index_style);
-write_file("index.html", html);
+}
