--- run.sh
@@ -1,1 +1,0 @@
-deno ./markblog.ts
